+++ lunchable/_cli.py
@@ -17,8 +17,6 @@ import lunchable
 from lunchable import LunchMoney
 from lunchable._config.logging_config import set_up_logging
 from lunchable.models import LunchableModel
-from lunchable.plugins.primelunch.primelunch import run_primelunch
-from lunchable.plugins.pushlunch import PushLunch
 
 logger = logging.getLogger(__name__)
 
@@ -381,6 +379,8 @@ def notify(continuous: bool, interval: int, user_key: str) -> None:
     """
     Send a Notification for each Uncleared Transaction
     """
+    from lunchable.plugins.pushlunch import PushLunch
+
     push = PushLunch(user_key=user_key)
     if interval is not None:
         interval = int(interval)
@@ -430,4 +430,51 @@ def primelunch() -> None:
     """
 
 
-primelunch.add_command(run_primelunch)
+@primelunch.command("run")
+@click.option(
+    "-f",
+    "--file",
+    "csv_file",
+    type=click.Path(exists=True, resolve_path=True),
+    help="File Path of the Amazon Export",
+    required=True,
+)
+@click.option(
+    "-w",
+    "--window",
+    "window",
+    type=click.INT,
+    help="Allowable time window between Amazon transaction date and "
+    "credit card transaction date",
+    default=7,
+)
+@click.option(
+    "-a",
+    "--all",
+    "update_all",
+    is_flag=True,
+    type=click.BOOL,
+    help="Whether to skip the confirmation step and simply update all matched "
+    "transactions",
+    default=False,
+)
+@click.option(
+    "-t",
+    "--token",
+    "access_token",
+    type=click.STRING,
+    help="LunchMoney Access Token - defaults to the LUNCHMONEY_ACCESS_TOKEN environment variable",
+    envvar="LUNCHMONEY_ACCESS_TOKEN",
+)
+def run_primelunch(
+    csv_file: str, window: int, update_all: bool, access_token: str
+) -> None:
+    """
+    Run the PrimeLunch Update Process
+    """
+    from lunchable.plugins.primelunch.primelunch import PrimeLunch
+
+    primelunch = PrimeLunch(
+        file_path=csv_file, time_window=window, access_token=access_token
+    )
+    primelunch.process_transactions(confirm=not update_all)
