+++ lunchable/plugins/primelunch/primelunch.py
@@ -11,14 +11,12 @@ import os
 import pathlib
 from typing import Any, Optional, Union
 
-import click
 import numpy as np
 import pandas as pd
 from numpy import datetime64
 from rich import print, table
 from rich.prompt import Confirm
 
-from lunchable._config.logging_config import set_up_logging
 from lunchable.models import (
     CategoriesObject,
     TransactionObject,
@@ -391,56 +389,3 @@ class PrimeLunch(LunchablePandasApp):
         else:
             float_string = f"[bold green]$ {float(amount):,.2f}[/bold green]"
         return float_string
-
-
-@click.command("run")
-@click.option(
-    "-f",
-    "--file",
-    "csv_file",
-    type=click.Path(exists=True, resolve_path=True),
-    help="File Path of the Amazon Export",
-    required=True,
-)
-@click.option(
-    "-w",
-    "--window",
-    "window",
-    type=click.INT,
-    help="Allowable time window between Amazon transaction date and "
-    "credit card transaction date",
-    default=7,
-)
-@click.option(
-    "-a",
-    "--all",
-    "update_all",
-    is_flag=True,
-    type=click.BOOL,
-    help="Whether to skip the confirmation step and simply update all matched "
-    "transactions",
-    default=False,
-)
-@click.option(
-    "-t",
-    "--token",
-    "access_token",
-    type=click.STRING,
-    help="LunchMoney Access Token - defaults to the LUNCHMONEY_ACCESS_TOKEN environment variable",
-    envvar="LUNCHMONEY_ACCESS_TOKEN",
-)
-def run_primelunch(
-    csv_file: str, window: int, update_all: bool, access_token: str
-) -> None:
-    """
-    Run the PrimeLunch Update Process
-    """
-    primelunch = PrimeLunch(
-        file_path=csv_file, time_window=window, access_token=access_token
-    )
-    primelunch.process_transactions(confirm=not update_all)
-
-
-if __name__ == "__main__":
-    set_up_logging(log_level=logging.INFO)
-    run_primelunch()
